backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.120.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
multidict==6.7.0
packaging==23.2
postgrest==2.22.2
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
realtime==2.22.2
sniffio==1.3.1
starlette==0.48.0
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.9.5
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
httpx==0.27.0
pyjwt==2.8.0
email-validator==2.1.0
